+++ docs/misc/changelog.rst
@@ -4,7 +4,7 @@ Changelog
 ==========
 
 
-Release 2.8.0a2 (WIP)
+Release 2.8.0a3 (WIP)
 --------------------------
 
 Breaking Changes:
@@ -19,6 +19,7 @@ New Features:
 Bug Fixes:
 ^^^^^^^^^^
 - Fixed saving and loading of Torch compiled models (using ``th.compile()``) by updating ``get_parameters()``
+- Added a warning to env-checker if a multidiscrete space has multi-dimensional array (@unexploredtest)
 
 `SB3-Contrib`_
 ^^^^^^^^^^^^^^
@@ -40,9 +41,12 @@ Others:
 - Improved type hints for distributions
 - Simplified zip file loading by removing Python 3.6 workaround and enabling ``weights_only=True`` (PyTorch 2.x)
 - Sped up saving/loading tests
+- Updated black from v25 to v26
+
 
 Documentation:
 ^^^^^^^^^^^^^^
+- Added a note on MultiDiscrete spaces with multi-dimensional arrays and a wrapper to fix the issue (@unexploredtest)
 
 
 Release 2.7.1 (2025-12-05)